refactor(yeoman): Generalize Yeoman-related code
Complete the core refactor of code to allow for a more generalized
approach to building Yeoman Generators inside SFDX-Falcon. This new code
should dramatically reduce the amount of repeated, boilerplate code when
building Yeoman Generators. This code also wires-in the SfdxFalconResult
object between Command and Generator, as well as uses the latest error
display code.

Related to: #22 #24