refactor(yeoman): Modify Falcon Prompt & Interview
Changed the prompt() method to return the User Answers instead of the
Final Answers.  This was done because, when used in the context of an
Interview, the SfdxFalconPrompt object would end up overwriting ALL user
answers in the interview because it was sending back final answers.
Also added better debug to SfdxFalconInterview.start()