refactor(yeoman): Update git tasks to use skips
For the "finalize Git" tasks, have all tasks after the first Git
executable check look for success/failure of each previous step and
SKIP themselves if the previous step was unsuccessful.  This required
that we set the "exitOnError" option to FALSE for the Git operations
task group, then process the outcome of the task run as a possible
exception.